local.conf.sample: Hold version for wpa_supplicant
The rockchip-rk3288-tci device type is using a 4.4 kernel that
does not support the MACset interface introduced in kernel v4.6
and required for WPA supplicant above 2.9.

Other Dunfell device types have newer kernels, so this is not
a Dunfell wide fix that could be set in the meta-balena integration
layer.

Also, the machine file for rockchip-rk3288-tci is in the vendor's
BSP layer, so there is no better place to set this preferred
version than local.conf.

Changelog-entry: Hold wpa supplicant version for rockchip-rk3288-tci
Signed-off-by: Alex Gonzalez <alexg@balena.io>